Move PolygonBuilder into Mapzen namespace
